+++ src/constants/default-datanautics-options.ts
@@ -2,12 +2,9 @@ import * as console from 'console';
 import { resolve } from 'path';
 
 import { DatanauticsOptions } from '@options';
-import { DEFAULT_DUMP_INTERVAL } from './default-dump-interval';
 
 export const defaultDatanauticsOptions: DatanauticsOptions = {
-  dumpInterval: DEFAULT_DUMP_INTERVAL,
   verbose: true,
   logger: console,
   dumpPath: resolve(__dirname, '../../data/data.txt'),
-  mode: 'writer'
 };