+++ src/constants/dump-event.ts
@@ -1,2 +1,0 @@
-export const WRITE_EVENT: string = 'write';
-export const READ_EVENT: string = 'read';