+++ src/constants/index.ts
@@ -1,3 +1,1 @@
-export * from './default-dump-interval';
-export * from './dump-event';
 export * from './default-datanautics-options';