+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] - 2025-06-20 (**Breaking changes!**)
+
+### Removed
+
+- Replaced Auto-Save Mechanism with Event Based Save Mechanism (triggers once you change data);
+- Removed **mode** and **dumpInterval** from options
+
+---
+
+
 ## [2.2.9] - 2025-06-20
 
 ### Added